+++ tes-rest-integrations/src/main/java/dk/sunepoulsen/tes/rest/integrations/TechEasySolutionsBackendIntegrator.java
@@ -1,7 +1,9 @@
 package dk.sunepoulsen.tes.rest.integrations;
 
+import dk.sunepoulsen.tes.rest.models.HashMapModel;
 import dk.sunepoulsen.tes.rest.models.monitoring.ServiceHealth;
 import dk.sunepoulsen.tes.rest.models.monitoring.ServiceInfo;
+import dk.sunepoulsen.tes.rest.models.monitoring.ServiceMetrics;
 import io.reactivex.rxjava3.core.Single;
 
 public class TechEasySolutionsBackendIntegrator extends AbstractIntegrator {
@@ -16,6 +18,21 @@ public class TechEasySolutionsBackendIntegrator extends AbstractIntegrator {
 
     public Single<ServiceInfo> info() {
         return Single.fromFuture(httpClient.get("/actuator/info", ServiceInfo.class))
+            .onErrorResumeNext(this::mapClientExceptions);
+    }
+
+    public Single<HashMapModel> env() {
+        return Single.fromFuture(httpClient.get("/actuator/env", HashMapModel.class))
+            .onErrorResumeNext(this::mapClientExceptions);
+    }
+
+    public Single<ServiceMetrics> metrics() {
+        return Single.fromFuture(httpClient.get("/actuator/metrics", ServiceMetrics.class))
+            .onErrorResumeNext(this::mapClientExceptions);
+    }
+
+    public Single<HashMapModel> metric(String name) {
+        return Single.fromFuture(httpClient.get("/actuator/metrics/" + name, HashMapModel.class))
             .onErrorResumeNext(this::mapClientExceptions);
     }